% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_shiny}
\alias{check_shiny}
\title{check shiny}
\usage{
check_shiny(path = here::here("R"))
}
\arguments{
\item{path}{path which stores the scripts}
}
\description{
This is designed with the  {golem} framework in mind, in that case the function
can be run without arguments in the relevant project. For optimal results attach
all packages used by the app and load all defined functions, as shinycheck doesn't
comment on functions that don't exist in the workspace.
}
